fix: problems on endpoints

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -4,6 +4,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.exception.FilmDateValidationException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.service.FilmService;
@@ -36,41 +37,52 @@ public class FilmController {
 
     @GetMapping("/films/{id}")
     public Film findFilmById(@PathVariable Long id){
+        if (id < 0) {
+            throw new ValidationException("id must be positive");
+        }
         return filmStorage.findById(id);
     }
 
     @PostMapping("/films")
-    public void create(@Validated @RequestBody Film film) {
+    public Film create(@Validated @RequestBody Film film) {
         log.info("Получен запрос к эндпоинту: POST /films, Строка параметров запроса: '{}'",
                 film.toString());
         validateFilm(film);
         filmStorage.create(film);
+        return film;
     }
 
     @PutMapping("/films")
-    public void update(@Validated @RequestBody Film film) {
+    public Film update(@Validated @RequestBody Film film) {
         log.info("Получен запрос к эндпоинту: PUT /films, Строка параметров запроса: '{}'",
                 film.toString());
         validateFilm(film);
         filmStorage.update(film);
+        return film;
     }
 
     private void validateFilm(Film film) {
         if (film.getReleaseDate().isBefore(LocalDate.of(1895, 12, 28))) {
-            throw new ValidationException("Era of films starts 28Dec1895 ");
+            throw new FilmDateValidationException("Era of films starts 28Dec1895 ");
         }
         if (film.getDuration().isNegative()) {
-            throw new ValidationException("Duration of film can't be negative");
+            throw new FilmDateValidationException("Duration of film can't be negative");
         }
     }
 
     @PutMapping("/films/{id}/like/{userId}")
     public void addLike(@PathVariable Long id, @PathVariable Long userId){
+        if (id < 0 || userId < 0) {
+            throw new ValidationException("id must be positive");
+        }
         filmService.addLike(id,userId);
     }
 
     @DeleteMapping("/films/{id}/like/{userId}")
     public  void deleteLike(@PathVariable Long id, @PathVariable Long userId){
+        if (id < 0 || userId < 0) {
+            throw new ValidationException("id must be positive");
+        }
         filmService.removeLike(id,userId);
     }
